# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: PaySurity API
  version: 2026.1.0
  description: Multi-tenant payment & verticals platform
servers:
  - url: https://paysurity-api.us-central1.run.app
  - url: http://localhost:3000
security:
  - BearerAuth: []
paths:
...
  /wallets/{walletId}/transfer:
    post:
      summary: Transfer funds between wallets
      operationId: transferFunds
      security:
        - BearerAuth: []
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - currency
                - recipientWalletId
              properties:
                amount:
                  type: number
                  format: float
                  minimum: 0.01
                  example: 50.00
                currency:
                  type: string
                  example: USD
                recipientWalletId:
                  type: string
                  format: uuid
                  example: 9z8y7x6w-5v4u-3t2s-1r0q-9p8o7n6m5l4k
                description:
                  type: string
                  nullable: true
                  example: Monthly allowance
      responses:
        '200':
          description: Funds transferred successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  senderTransactionId:
                    type: string
                    format: uuid
                    example: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                  recipientTransactionId:
                    type: string
                    format: uuid
                    example: f0e9d8c7-b6a5-4321-fedc-ba9876543210
                  senderNewBalance:
                    type: number
                    format: float
                    example: 1550.75
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - statusCode
        - message
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Bad Request
        details:
          type: array
          items:
            type: string
          nullable: true
          example: ["Validation failed: email must be an email"]
  responses:
    BadRequest:
      description: Bad Request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
            example:
              statusCode: 400
              message: "Invalid input"
              details: ["Field 'amount' is required."]
    Unauthorized:
      description: Unauthorized - Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
            example:
              statusCode: 401
              message: "Unauthorized"
              details: ["Invalid or expired access token."]
    Forbidden:
      description: Forbidden - User does not have necessary permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
            example:
              statusCode: 403
              message: "Forbidden"
              details: ["User does not have access to this resource."]
    NotFound:
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
            example:
              statusCode: 404
              message: "Resource not found"
              details: ["Wallet with ID '...' not found."]
    InternalServerError:
      description: Internal Server Error - Something went wrong on the server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
            example:
              statusCode: 500
              message: "Internal Server Error"
              details: ["An unexpected error occurred. Please try again later."]